requirements.txt
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
reportlab==4.4.2
